backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "numpy>=2.4.3",
    "openai>=1.75.0",
    "python-dotenv>=1.0.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.48",
    "mysql-connector-python>=9.6.0",
    "requests>=2.33.0",
]
